refactor: application 계층 service 계층으로 변경
